* Fix some todo, replace ASSERT of hardware part by ASSERT_KP

+++ code/userprog/stackmanager.h
@@ -4,13 +4,12 @@
 class AddrSpace;
 class Lock;
 
-// TODO: Regroup flags into a bitfield
 struct StackRegion {
     unsigned int base;      // High address (top of stack, where SP starts)
     unsigned int limit;     // Low address (bottom of stack)
     bool inUse;             // Currently allocated to a thread
     bool userProvided;      // Stack was provided by user (don't free memory)
-    unsigned int threadTid; // TID of thread using this stack (for debugging)
+    unsigned short threadTid; // TID of thread using this stack (for debugging)
 
     StackRegion() : base(0), limit(0), inUse(false), userProvided(false), threadTid(0) {}
